Issue #24 - add help option to footer and as a menu item in 'My Account' options block.
The 'RiverRunEPerson' aspect created here *adds to* the existing DSpace EPerson aspect.
Since the additons are simple and always present, we use XSL-driven aspect modifications
to add the menu option to the 'My Account' block.

The help link in the footer is handled in the theme layer -- this is consistent with the
'Contact' and 'Feedback' links in the same section.